refactor[st]: removed not needed parts
