+++ SmartTutor/ContentModel/IChallengeService.cs
@@ -1,5 +1,4 @@
-﻿using RepositoryCompiler.CodeModel.CaDETModel.CodeItems;
-using SmartTutor.ContentModel.LearningObjects.ChallengeModel;
+﻿using SmartTutor.ContentModel.LearningObjects.ChallengeModel;
 using SmartTutor.ContentModel.LearningObjects.ChallengeModel.FulfillmentStrategy;
 using System.Collections.Generic;
 
@@ -10,6 +9,5 @@ namespace SmartTutor.ContentModel
         ChallengeEvaluation CheckChallengeCompletion(string[] sourceCode, int challengeId);
         Challenge GetChallenge(int challengeId);
         List<ChallengeHint> GetAllHints(int challengeId);
-        List<ChallengeHint> GetHintsForSolutionAttempt(int challengeId, List<CaDETClass> solutionAttempt);
     }
 }
