--- SmartTutor/Controllers/DTOs/Challenge/MetricRangeRuleDTO.cs
@@ -1,12 +1,0 @@
-﻿namespace SmartTutor.Controllers.DTOs.Challenge
-{
-    public class MetricRangeRuleDTO
-    {
-        public int Id { get; set; }
-        public string MetricName { get; set; }
-        public double FromValue { get; set; }
-        public double ToValue { get; set; }
-        public ChallengeHintDTO Hint { get; set; }
-
-    }
-}
