+++ SmartTutor/Controllers/DTOs/Lecture/ChallengeDTO.cs
@@ -1,11 +1,8 @@
-﻿using SmartTutor.Controllers.DTOs.Challenge;
-
-namespace SmartTutor.Controllers.DTOs.Lecture
+﻿namespace SmartTutor.Controllers.DTOs.Lecture
 {
     public class ChallengeDTO : LearningObjectDTO
     {
         public string Url { get; set; }
         public string Description { get; set; }
-        public ChallengeFulfillmentStrategyDTO FulfillmentStrategy { get; set; }
     }
 }
