+++ SmartTutorTests/Unit/ChallengeServiceTests.cs
@@ -1,7 +1,6 @@
 ﻿using Shouldly;
 using SmartTutor.ContentModel;
 using SmartTutor.ContentModel.LearningObjects.ChallengeModel;
-using SmartTutor.ContentModel.LearningObjects.ChallengeModel.FulfillmentStrategy.MetricChecker;
 using SmartTutor.ContentModel.LearningObjects.Repository;
 using Xunit;
 
@@ -24,27 +23,7 @@ namespace SmartTutorTests.Unit
             challenge.Id.ShouldBe(3371);
             challenge.LearningObjectSummaryId.ShouldBe(337);
             challenge.Url.ShouldBe("https://github.com/Ana00000/Challenge-inspiration.git");
-
-            BasicMetricsChecker basicMetricsChecker = (BasicMetricsChecker)challenge.FulfillmentStrategy;
-            basicMetricsChecker.ClassMetricRules.Count.ShouldBe(2);
-            basicMetricsChecker.ClassMetricRules[0].Id.ShouldBe(33701);
-            basicMetricsChecker.ClassMetricRules[0].MetricName.ShouldBe("CLOC");
-            basicMetricsChecker.ClassMetricRules[0].FromValue.ShouldBe(3);
-            basicMetricsChecker.ClassMetricRules[0].ToValue.ShouldBe(30);
-            basicMetricsChecker.ClassMetricRules[1].Id.ShouldBe(33702);
-            basicMetricsChecker.ClassMetricRules[1].MetricName.ShouldBe("NMD");
-            basicMetricsChecker.ClassMetricRules[1].FromValue.ShouldBe(0);
-            basicMetricsChecker.ClassMetricRules[1].ToValue.ShouldBe(2);
-            basicMetricsChecker.MethodMetricRules.Count.ShouldBe(2);
-            basicMetricsChecker.MethodMetricRules[0].Id.ShouldBe(33703);
-            basicMetricsChecker.MethodMetricRules[0].MetricName.ShouldBe("MELOC");
-            basicMetricsChecker.MethodMetricRules[0].FromValue.ShouldBe(2);
-            basicMetricsChecker.MethodMetricRules[0].ToValue.ShouldBe(5);
-            basicMetricsChecker.MethodMetricRules[1].Id.ShouldBe(33704);
-            basicMetricsChecker.MethodMetricRules[1].MetricName.ShouldBe("NOP");
-            basicMetricsChecker.MethodMetricRules[1].FromValue.ShouldBe(1);
-            basicMetricsChecker.MethodMetricRules[1].ToValue.ShouldBe(4);
-
+            challenge.FulfillmentStrategy.ShouldNotBe(null);
             challenge.FulfillmentStrategy.ChallengeHints.Count.ShouldBe(2);
             challenge.FulfillmentStrategy.ChallengeHints[0].Id.ShouldBe(337001);
             challenge.FulfillmentStrategy.ChallengeHints[0].Content.ShouldBe("Cohesion");
